requirements.txt
fastapi
uvicorn
pillow
qrcode
pandas